pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-personality-experiment"
version = "0.1.0"
description = "Experimental framework for studying LLM personalities with adaptive weighted selection."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "matplotlib>=3.8",
  "pydantic>=2.6",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
llm-personality-experiment = "llm_personality_experiment.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
